// sourcecode/BTrak.Packages/snova-testrepo/testrepo-module-components/testrepo-components/src/lib/snovasys-testrepo/models/GoalModel.ts
import { EntityRoleFeatureModel } from './entityRoleFeature';

export class GoalModel {
  goalId: string;
  goalName: string;
  goalShortName: string;
  goalStatusId: string;
  goalStatusName: string;
  goalResponsibleUserId: string;
  goalResponsiblePersonFirstName: string;
  goalResponsiblePersonSurName: string;
  goalResponsiblePerson: string;
  goalResponsiblePersonProfileImage: string;
  goalDeadLineDate: string;
  onboardProcessDate: Date;
  onboardDate: string;
  boardTypeId: string;
  boardTypeUiId: string;
  isArchived: boolean;
  archivedDateTime: any;
  createdDateTime: any;
  userStoriesCount: number;
  bugsCount: number;
  isWarning: boolean;
  dueDays: number;
  dueDaysCount: number;
  boardTypeApiId: string;
  maxDeadLine: string;
  goalStatusColor: string;
  goalsCount: number;
  canUserStoryPark: boolean;
  configurationId: string;
  isProductiveBoard: boolean;
  isToBeTracked: boolean;
  considerEstimatedHoursId: string;
  version: string;
  projectId: string;
  isParked: boolean;
  boardTypeName: string;
  profileImage: string;
  isApproved: boolean;
  isLocked: boolean;
  totalCount: number;
  activeGoalsCount: number;
  backLogGoalsCount: number;
  underReplanGoalsCount: number;
  goalArchivedDateTime: Date;
  activeUserStoryCount: number;
  parkedDateTime: Date;
  goalResponsibleUserName: string;
  goalResponsibleProfileImage: string;
  goalDeadLine: Date;
  goalEstimatedTime: any;
  workflowId: string;
  entityFeaturesXml: string;
  entityFeaturesList: EntityRoleFeatureModel[];
  timeStamp: any;
  isMovingToAnotherProject: boolean;
  isEdit: boolean;
  inActiveDateTime: Date;
  testSuiteId: string;
  isBugBoard: boolean;
  description : string;
  tag : string;
  updatedDateTime: any;
  goalLabel: string;
  isDateTimeConfiguration: boolean;
  isSprintsConfiguration: boolean;
  isEnableTestRepo: boolean;
}