// sourcecode/BTrak.Packages/snova-testrepo/testrepo-module-components/testrepo-components/src/lib/snovasys-testrepo/models/sprints-model.ts
export class SprintModel {
    sprintId: string;
    projectId: string;
    sprintName: string;
    projectName: string;
    sprintStartDate: Date;
    sprintEndDate: Date;
    createdDateTime: any;
    createdUserName: string;
    inActiveDateTime: Date;
    isBacklog: boolean;
    isArchived: boolean;
    timeStamp: any;
    pageNumber: number;
    pageSize: number;
    totalCount: number;
    description: string;
    isRefresh: boolean;
    isReplan: boolean;
    boardTypeId: string;
    boardTypeApiId: string;
    version: string;
    testSuiteId: string;
    isBugBoard: boolean;
    boardTypeUiId: string;
    workFlowId: string;
    isWarning: boolean;
    userStoriesCount: number;
    completedUserStoriesCount: number;
    sprintIds: string;
    sprintUniqueName: string;
    sprintResponsiblePersonName: string;
    profileImage: string;
    isComplete: boolean;
    editSprint: boolean;
    sprintResponsiblePersonId: string;
    isEnableSprints: boolean;
    isEnableTestRepo: boolean;
}